1.4.8.8
- station recipes tinkered with to make progression easier. This
includes the quantum extractor.
- new armor: War Angel. Essentially the same as Elder in terms of
protective effects, but the armor is slightly higher, as is the energy
and it has very different perma-buffs applied to it. It’s also a son of
a bitch to make. Unlocks with Nuclear Core. Enjoy.
- vanilla moons should now spawn moonstone
- fixed white background on elder altar icon
- fixed missing unlock for obsidian shard and magnetic shielding
- fixed proto-chest recipe component
- changed darkstone requirement in stone chests to moonrock
- irradiated tiles now produce half as much uranium when extracted. It
was a bit absurd at 2 per 50, given that uranium is already quite common
in irradiated biomes