huuuuge extractor change
New Extractor code in play. You will need to add the old sampling array
with /spawnitem samplingarray  and place it that way. It is using the
code at the moment (a modified version of the old array)

other than that, we have:

changes to tier 4 mining gun projectile and weapon appearance

new mining gun (tier 5)