biome improvements, bug fixes
- Removed a few instances of the “ledges” terrain generator. It is now
confined exclusively to less friendly worlds where you are likely to see
huge cliffs occurring naturally (such as Ice Wastes , Sulphuric worlds
and the like). You will not see them on Gelatinous worlds any more, save
for specific sub-biomes on the surface.
- replaced existing Ledges terrain generator with a “new” one (read:
lowered height values substantially). Hopefully this solves the
“giant-assed wall of tiles out of nowhere” issue :)
- added new terrain generators
- tarstone should now properly extract
- junglestone should now properly extract
- added molten iron liquid. It inflicts heat-based damage that can only
be blocked with Lava immunity effects
- molten iron now falls as a rain type on Metallic Moons, and also
appears sporadically as a ground-liquid type on that biome, and a few
secondary biomes.
- liquid mercury now inflicts less damage per tick
- pekkitbomb freeze time increased substantially, and it has its own
“effect” now, which makes it a bit more visually blue when it freezes
foes.
- minor primeval forest parallax adjustments
- fixed treefront/treeback parallax issue on crystalmoons
- replaced corrupted healingrain image with a proper one. it was blank
either way, but the error message was annoying me.
- removed references to “humanencounter” as a minidungeon since it no
longer exists in the vanilla files.