microdungeon additions/fixes
-added 3 new blocks: Titanium Wall, Speed Block and Jump Block. The last
two require Titanium Wall and other resources to build. They are
incredibly useful. Pick up a fission furnace to unlock the recipes.
-added FU original (thanks, @dasmole for the art and code!) fleshleech
monster to bloodstone and atropus
- heavily reduced FU monster spawn rates
-elderspear blueprint should hopefully work okay now, rather than being
a PGI
- added some new decorations to primeval forest worlds
- rostone should properly extract again
- you can now build the brain extractor once you find silicon, simply to
enable the recipe for Brain in a Jar to work properly
- dinos now have a chance to drop eggs, which can be extracted from for
materials
- added one new ice surface microdungeon
- added two brand new arboreal surface microdungeons. a giant tree, a
bluff,  and a dangerous bridge. It seems they spawn very rarely even
when kept as the only selection options in the biome set to max spawn
rate, but that is due to their size and that they require a fair amount
of background being strictly air.