shadow improvements
- shadow moons got some love. or evil. or something like that.
- along with the above, added a bunch of nightly stuff for
midnight/shadow biome. These will be removed when nightly goes to
Stable.
- grey metallic trees from vanilla now drop metal-coated wood instead of
ordinary wood (not sure why they made this change. Possibly
unintentional. Thanks to @Treason for pointing it out)
- more uses for Blood Diamonds. Mostly, I swapped out the need for
densealloy and replaced it with blooddiamond in the quietus weapon
recipes and replacing biochip with blooddiamond on the armor recipe.