big stuff!
- substantially weakened the absolutely LUDICROUSLY overpowered Swarm
guns. They now act balanced rather than one-hitting the Shoggoth! Go by
their Damage Per Shot more than their DPS, as they are a special-case
weapon.
- Extraction Lab now extracts at 0.95 second intervals rather than 1.12.
It’s a slight adjustment but you’ll likely notice the difference.
- Methanol now unlocks Methyl Iodide. This should make it a bit easier
to work out, as the only other unlocks were pesticide (which needs it in
the first place in order to craft) and Fertilizer 2.
- new look for Vashta Gun
- moved XT-3 Laser Cannon unlock to Eridian Ore
- Added Ansible Network station. You can use this to buy various goods
- some stupid-assed issue is making the steampunk helm blueprint appear
as a PGI in the cinnabar quest reward. So, I swapped it to the helm
itself (and the helm now teaches the recipe)
- new solar panel art (minor changes to make it suit the FU lab gear
better)

- new station: Solar Array : This is a bigger, more proficient version
of the small solar panel
- new station: Ansible Network : Provides ability to purchase goods from
your base. (no selling though, nor will this be added). To use, wire it
up to power. Then, with wiring tool and paint tool [one for forward, one
for back ] change “channels” to the shop you want. Then swap to your
hands to use the station itself. The station is still pretty bare with
its offerings. This will change.
-new station: Quantum Generator : Biggest and best of the generators.
The art is temporary (very), as the final will be substantially larger
and more interesting (and difficult to craft).