tonsa goodies
Stations
------------
- thermal generator no longer takes Solarium as a fuel source (since it
is not a fission-based reactor)
- recolored most of the Science mod stations to better match the FU
stations and tiles
- all vanilla Bugs extract for resources
- all bee Honeycombs now extract for resources
- Extractors can now convert toxic waste into uranium ore (10:1 basic,
5:1 quantum)
- fixed a few issues with the arc smelter. It no longer creates insane
amounts of diamond from Carbon, and some PGIs have been corrected
- Atmosphere Condenser now unlocks from both Battery2, and the Thermal
Generator.
- obsidian should (finally) extract properly. I kept forgetting about
it!
- Extraction Lab now extracts at 0.95 second intervals rather than 1.12.
It’s a slight adjustment but you’ll likely notice the difference.
- new solar panel art (minor changes to make it suit the FU lab gear
better)
- new station: Solar Array : This is a bigger, more proficient version
of the small solar panel
- new station: Ansible Network : Provides ability to purchase goods from
your base. (no selling though, nor will this be added). To use, wire it
up to power. Then, with wiring tool and paint tool [one for forward, one
for back ] change “channels” to the shop you want. Then swap to your
hands to use the station itself. The station is still pretty bare with
its offerings. This will change.
-new station: Quantum Generator : Biggest and best of the generators.
The art is temporary (very), as the final will be substantially larger
and more interesting (and difficult to craft).
- new station:  Centrifuge : (old “autoclave” has been removed and
replaced with this). You can set liquids into it and get goods back.

Biomes
------------
- sulphuric biome weather does a bit more damage. It wasn’t quite
dangerous enough.
- added Contaminated Water to Bog and Wasteland biomes
- added regular water to Bog biomes

Dungeons
---------------
- FU wreck dungeon lockers now contain basic loot
- loot changes on several other dungeons (swapped from basic treasure to
FU basic treasure)

Gear
---------------
- new look for Vashta Gun
- substantially weakened the absolutely LUDICROUSLY overpowered Swarm
guns. They now act balanced rather than one-hitting the Shoggoth! Go by
their Damage Per Shot more than their DPS, as they are a special-case
weapon.
- moved XT-3 Laser Cannon unlock to Eridian Ore

Quests
------------
- some stupid-assed issue is making the steampunk helm blueprint appear
as a PGI in the cinnabar quest reward. So, I swapped it to the helm
itself (and the helm now teaches the recipe)

Monsters
---------------

Resources
-----------------
- added Contaminated Water
- most (not all) liquids can now be crafted using Contaminated Water and
additional materials.
- Methanol now unlocks Methyl Iodide. This should make it a bit easier
to work out, as the only other unlocks were pesticide (which needs it in
the first place in order to craft) and Fertilizer 2.
- added Enriched Uranium, Enriched Plutonium and Ultronium. These can
only be obtained via the Centrifuge and are useful as fuel for various
powered stations or your ship.